+++ move.go
@@ -1,6 +1,6 @@
 package main
 
-func getValuePostion(board [][]int, toFind int) Pos2D {
+func getValuePostion(board [][]uint8, toFind uint8) Pos2D {
 	for i, line := range board {
 		for j, value := range line {
 			if value == toFind {
@@ -15,7 +15,7 @@ func swap[T any](a, b *T) {
 	*a, *b = *b, *a
 }
 
-func moveUp(board [][] int) (ok bool, updatedBoard [][]int) {
+func moveUp(board [][] uint8) (ok bool, updatedBoard [][]uint8) {
 	empty := getValuePostion(board, 0)
 	updatedBoard = Deep2DSliceCopy(board)
 	if empty.Y != 0 {
@@ -25,7 +25,7 @@ func moveUp(board [][] int) (ok bool, updatedBoard [][]int) {
 	return false, nil
 }
 
-func moveDown(board [][]int) (ok bool, updatedBoard [][]int) {
+func moveDown(board [][]uint8) (ok bool, updatedBoard [][]uint8) {
 	empty := getValuePostion(board, 0)
 	updatedBoard = Deep2DSliceCopy(board)
 	if empty.Y != len(board)-1 {
@@ -35,7 +35,7 @@ func moveDown(board [][]int) (ok bool, updatedBoard [][]int) {
 	return false, nil
 }
 
-func moveLeft(board [][]int) (ok bool, updatedBoard [][]int) {
+func moveLeft(board [][]uint8) (ok bool, updatedBoard [][]uint8) {
 	empty := getValuePostion(board, 0)
 	updatedBoard = Deep2DSliceCopy(board)
 	if empty.X != 0 {
@@ -45,7 +45,7 @@ func moveLeft(board [][]int) (ok bool, updatedBoard [][]int) {
 	return false, nil
 }
 
-func moveRight(board [][]int) (ok bool, updatedBoard [][]int) {
+func moveRight(board [][]uint8) (ok bool, updatedBoard [][]uint8) {
 	empty := getValuePostion(board, 0)
 	updatedBoard = Deep2DSliceCopy(board)
 	if empty.X != len(board)-1 {
